# api/openapi.yaml
...
openapi: 3.0.0
info:
  title: A2OS SafeU
  description: |
    A2OS SafeU 项目 API 文档。
  contact:
    email: a2os-general@googlegroups.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0.0-alpha"
...
paths:
...
  /v1/upload/finish:
    post:
      tags:
        - upload
      summary: 结束所有上传文件的上传操作
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    example: fe5305b4-e0c8-490b-b3dc-eb523e699a07
                  description: 文件列表
      responses:
        "200":
          description: 结束文件上传成功
          content:
            "application/json":
              schema:
                type: object
                properties:
                  owner:
                    type: string
                    description: 文件所有者信息
                  recode:
                    type: string
                    description: 文件提取码
...
  /v1/item/{retrieveCode}:
    post:
      tags:
        - download
      summary: 获取下载/打包文件连接
      operationId: downloadItem
      parameters:
        - name: Token
          in: header
          description: 临时认证口令
          style: simple
          required: true
          schema:
            type: string
        - name: retrieveCode
          in: path
          description: 提取码
          required: true
          schema:
            type: string
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/PackRequest"
      responses:
        "200":
          description: 成功获取文件（组）下载链接
          content:
            "application/json":
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: 文件（组）下载链接
        "400":
          description: 无法获得文件列表
          content:
            "application/json":
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: 错误原因
        "401":
          description: 认证失败
          content:
            "application/json":
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: 错误原因
        "404":
          description: 无法找到资源/认证口令
          content:
            "application/json":
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: 错误原因
        "410":
          description: 下载次数耗尽
          content:
            "application/json":
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: 错误原因
        "503":
          description: 无法提供服务
          content:
            "application/json":
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: 错误原因
...
components:
  schemas:
...
    ZipItem:
      type: object
      description: 待下载文件
      properties:
        protocol:
          type: string
          description: OSS 下载协议
        bucket:
          type: string
          description: 文件所在 OSS 桶名
        endpoint:
          type: string
          description: OSS 地域路径
        path:
          type: string
          description: 文件所在 OSS 桶中的绝对路径
        original_name:
          type: string
          description: 文件原名
    PackRequest:
      type: object
      description: 打包/下载链接请求
      properties:
        full:
          type: boolean
          description: 是否为该提取码下的全部文件
        items:
          type: array
          description: 待打包/下载文件
          items:
            $ref: "#/components/schemas/ZipItem"